% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fringeColValidations.R
\name{fringeColVal_greaterThan0}
\alias{fringeColVal_greaterThan0}
\title{fringeColVal_greaterThan0}
\usage{
fringeColVal_greaterThan0(fringe, cols)
}
\description{
fringeColVal_greaterThan0
WORKS_WITH_CTYPE: N
}
