% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fringeColValidations.R
\name{fringeColVal_hasGenderLevelsEs}
\alias{fringeColVal_hasGenderLevelsEs}
\title{fringeColVal_hasGenderLevelsEs}
\usage{
fringeColVal_hasGenderLevelsEs(fringe, cols)
}
\description{
fringeColVal_hasGenderLevelsEs
}

