% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sysfile}
\alias{sysfile}
\title{Read contents of a system file into a character string}
\usage{
sysfile(..., package = "datafringe")
}
\arguments{
\item{string}{string}
}
\value{
string
}
\description{
sysfile
}
\examples{
\dontrun{
}
}
